mv some common fns
