+++ typescript/infra/src/agents/aws/user.ts
@@ -7,7 +7,11 @@ import {
 } from '@aws-sdk/client-iam';
 import { KEY_ROLE_ENUM } from '../../agents';
 import { AgentConfig } from '../../config';
-import { fetchGCPSecret, gcpSecretExists, setGCPSecret } from '../../utils/gcloud';
+import {
+  fetchGCPSecret,
+  gcpSecretExists,
+  setGCPSecret,
+} from '../../utils/gcloud';
 import { AgentAwsKey } from './key';
 
 export class AgentAwsUser<Networks extends ChainName> {
@@ -63,15 +67,13 @@ export class AgentAwsUser<Networks extends ChainName> {
   }
 
   async getAccessKeys(): Promise<{
-    accessKeyId: string,
-    secretAccessKey: string,
+    accessKeyId: string;
+    secretAccessKey: string;
   }> {
     return {
       accessKeyId: await fetchGCPSecret(this.accessKeyIdSecretName),
-      secretAccessKey: await fetchGCPSecret(
-        this.secretAccessKeySecretName,
-      ),
-    }
+      secretAccessKey: await fetchGCPSecret(this.secretAccessKeySecretName),
+    };
   }
 
   async createAndSaveAccessKey() {
